Moved RenderTarget to Graphics.Shared
